Rename modules to lower_snake_case
