+++ updater/site/wowace.py
@@ -1,6 +1,6 @@
 import requests
 
-from updater.site.AbstractSite import AbstractSite
+from updater.site.abstract_site import AbstractSite
 
 
 class WoWAce(AbstractSite):
